Ripping more stuff out.
