+++ datasource/api/src/main/java/org/opennms/oce/datasource/api/AlarmDatasource.java
@@ -26,29 +26,18 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.oce.engine.shell;
+package org.opennms.oce.datasource.api;
 
 import java.util.List;
 
-import org.apache.karaf.shell.api.action.lifecycle.Reference;
-import org.apache.karaf.shell.api.action.lifecycle.Service;
-import org.apache.karaf.shell.api.console.CommandLine;
-import org.apache.karaf.shell.api.console.Completer;
-import org.apache.karaf.shell.api.console.Session;
-import org.apache.karaf.shell.support.completers.StringsCompleter;
-import api.ScoringStrategy;
+public interface AlarmDatasource {
 
-@Service
-public class ScoreNameCompleter implements Completer {
+    List<? extends Alarm> getAlarms();
 
-    @Reference
-    private List<ScoringStrategy> strategies;
+    List<? extends Alarm> getAlarmsAndRegisterHandler(AlarmHandler handler);
 
-    @Override
-    public int complete(Session session, CommandLine commandLine, List<String> candidates) {
-        StringsCompleter delegate = new StringsCompleter();
-        strategies.forEach(s -> delegate.getStrings().add(s.getName()));
-        return delegate.complete(session, commandLine, candidates);
-    }
+    void registerHandler(AlarmHandler handler);
+
+    void unregisterHandler(AlarmHandler handler);
 
 }
