+++ datasource/common/src/main/java/org/opennms/oce/datasource/common/StaticAlarmDatasource.java
@@ -26,29 +26,39 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.oce.engine.shell;
+package org.opennms.oce.datasource.common;
 
 import java.util.List;
+import java.util.Objects;
 
-import org.apache.karaf.shell.api.action.lifecycle.Reference;
-import org.apache.karaf.shell.api.action.lifecycle.Service;
-import org.apache.karaf.shell.api.console.CommandLine;
-import org.apache.karaf.shell.api.console.Completer;
-import org.apache.karaf.shell.api.console.Session;
-import org.apache.karaf.shell.support.completers.StringsCompleter;
-import api.ScoringStrategy;
+import org.opennms.oce.datasource.api.Alarm;
+import org.opennms.oce.datasource.api.AlarmDatasource;
+import org.opennms.oce.datasource.api.AlarmHandler;
 
-@Service
-public class ScoreNameCompleter implements Completer {
+public class StaticAlarmDatasource implements AlarmDatasource {
+    private final List<? extends Alarm> alarms;
 
-    @Reference
-    private List<ScoringStrategy> strategies;
-
-    @Override
-    public int complete(Session session, CommandLine commandLine, List<String> candidates) {
-        StringsCompleter delegate = new StringsCompleter();
-        strategies.forEach(s -> delegate.getStrings().add(s.getName()));
-        return delegate.complete(session, commandLine, candidates);
+    public StaticAlarmDatasource(List<? extends Alarm> alarms) {
+        this.alarms = Objects.requireNonNull(alarms);
     }
 
+    @Override
+    public List<? extends Alarm> getAlarms() {
+        return alarms;
+    }
+
+    @Override
+    public List<? extends Alarm> getAlarmsAndRegisterHandler(AlarmHandler handler) {
+        return alarms;
+    }
+
+    @Override
+    public void registerHandler(AlarmHandler handler) {
+        // pass
+    }
+
+    @Override
+    public void unregisterHandler(AlarmHandler handler) {
+        // pass
+    }
 }
