+++ datasource/jaxb/pom.xml
@@ -6,7 +6,7 @@
         <version>1.0.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-    <groupId>org.opennms.oce.datasource.jaxb</groupId>
+    <groupId>org.opennms.oce.datasource</groupId>
     <artifactId>org.opennms.oce.datasource.jaxb</artifactId>
     <name>OCE :: Datasource :: JAXB</name>
     <packaging>bundle</packaging>
@@ -70,6 +70,16 @@
         <dependency>
             <groupId>org.opennms.oce.datasource</groupId>
             <artifactId>org.opennms.oce.datasource.common</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.osgi</groupId>
+            <artifactId>org.osgi.core</artifactId>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.osgi</groupId>
+            <artifactId>org.osgi.compendium</artifactId>
+            <scope>provided</scope>
         </dependency>
         <dependency>
             <groupId>org.jvnet.jaxb2_commons</groupId>
