+++ datasource/opennms/src/main/java/org/opennms/oce/datasource/opennms/OpennmsInventoryDatasource.java
@@ -26,29 +26,16 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.oce.engine.shell;
+package org.opennms.oce.datasource.opennms;
 
-import java.util.List;
+import java.util.Collections;
 
-import org.apache.karaf.shell.api.action.lifecycle.Reference;
-import org.apache.karaf.shell.api.action.lifecycle.Service;
-import org.apache.karaf.shell.api.console.CommandLine;
-import org.apache.karaf.shell.api.console.Completer;
-import org.apache.karaf.shell.api.console.Session;
-import org.apache.karaf.shell.support.completers.StringsCompleter;
-import api.ScoringStrategy;
+import org.opennms.oce.datasource.common.StaticInventoryDatasource;
 
-@Service
-public class ScoreNameCompleter implements Completer {
+public class OpennmsInventoryDatasource extends StaticInventoryDatasource {
 
-    @Reference
-    private List<ScoringStrategy> strategies;
-
-    @Override
-    public int complete(Session session, CommandLine commandLine, List<String> candidates) {
-        StringsCompleter delegate = new StringsCompleter();
-        strategies.forEach(s -> delegate.getStrings().add(s.getName()));
-        return delegate.complete(session, commandLine, candidates);
+    public OpennmsInventoryDatasource() {
+        super(Collections.emptyList());
     }
 
 }
