+++ datasource/opennms/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,0 +1,40 @@
+<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
+           xmlns:cm="http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.1.0"
+           xmlns:ext="http://aries.apache.org/blueprint/xmlns/blueprint-ext/v1.0.0">
+
+    <!-- Configuration properties -->
+    <cm:property-placeholder id="datasourceProperties" persistent-id="org.opennms.oce.datasource.opennms" update-strategy="reload">
+        <cm:default-properties>
+            <cm:property name="eventTopic" value="events"/>
+            <cm:property name="alarmTopic" value="alarms"/>
+            <cm:property name="nodeTopic" value="nodes"/>
+            <cm:property name="url" value="http://127.0.0.1:8980/opennms"/>
+            <cm:property name="username" value="admin"/>
+            <cm:property name="password" value="admin"/>
+        </cm:default-properties>
+    </cm:property-placeholder>
+
+    <reference id="configAdmin" interface="org.osgi.service.cm.ConfigurationAdmin" />
+    <bean id="alarmDatasource" class="org.opennms.oce.datasource.opennms.OpennmsAlarmDatasource" init-method="init" destroy-method="destroy">
+        <argument ref="configAdmin"/>
+        <property name="eventTopic" value="${eventTopic}"/>
+        <property name="alarmTopic" value="${alarmTopic}"/>
+        <property name="nodeTopic" value="${nodeTopic}"/>
+    </bean>
+    <service ref="alarmDatasource" interface="org.opennms.oce.datasource.api.AlarmDatasource"/>
+
+    <bean id="inventoryDatasource" class="org.opennms.oce.datasource.opennms.OpennmsInventoryDatasource"/>
+    <service ref="inventoryDatasource" interface="org.opennms.oce.datasource.api.InventoryDatasource"/>
+
+    <bean id="restClient" class="org.opennms.oce.connector.impl.OpennmsRestClient">
+        <argument value="${url}"/>
+        <argument value="${username}"/>
+        <argument value="${password}"/>
+    </bean>
+
+    <bean id="incidentDatasource" class="org.opennms.oce.datasource.opennms.OpennmsIncidentDatasource">
+        <argument ref="restClient"/>
+    </bean>
+    <service ref="incidentDatasource" interface="org.opennms.oce.datasource.api.IncidentDatasource"/>
+
+</blueprint>
