+++ driver/main/src/main/java/org/opennms/oce/driver/score/impl/ScoreMetricBean.java
@@ -26,29 +26,59 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.oce.engine.shell;
+package org.opennms.oce.driver.score.impl;
 
-import java.util.List;
+import org.opennms.oce.driver.score.api.ScoreMetric;
 
-import org.apache.karaf.shell.api.action.lifecycle.Reference;
-import org.apache.karaf.shell.api.action.lifecycle.Service;
-import org.apache.karaf.shell.api.console.CommandLine;
-import org.apache.karaf.shell.api.console.Completer;
-import org.apache.karaf.shell.api.console.Session;
-import org.apache.karaf.shell.support.completers.StringsCompleter;
-import api.ScoringStrategy;
+public class ScoreMetricBean implements ScoreMetric {
+    private String name;
+    private double value;
+    private String description;
 
-@Service
-public class ScoreNameCompleter implements Completer {
+    public ScoreMetricBean() {
 
-    @Reference
-    private List<ScoringStrategy> strategies;
-
-    @Override
-    public int complete(Session session, CommandLine commandLine, List<String> candidates) {
-        StringsCompleter delegate = new StringsCompleter();
-        strategies.forEach(s -> delegate.getStrings().add(s.getName()));
-        return delegate.complete(session, commandLine, candidates);
     }
 
+    public ScoreMetricBean(String name, String description) {
+        this.name = name;
+        this.description = description;
+    }
+
+    public ScoreMetricBean(String name, double value, String description) {
+        this.name = name;
+        this.value = value;
+        this.description = description;
+    }
+
+    @Override
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    @Override
+    public double getValue() {
+        return value;
+    }
+
+    public void setValue(double value) {
+        this.value = value;
+    }
+
+    @Override
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    @Override
+    public String toString() {
+        return name + ":" + Double.valueOf(value);
+    }
 }
