--- engine/driver/src/main/java/org/opennms/oce/engine/score/api/ScoringStrategy.java
@@ -1,50 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2018 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2018 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.oce.engine.score.api;
-
-import java.util.Set;
-
-import org.opennms.oce.model.alarm.api.Incident;
-
-public interface ScoringStrategy {
-
-    /**
-     * Issue a ScoreReport for a Set of Incitdents against a baseline.
-     * @param baseline The baseline set of Incidents
-     * @param sut The Set Under Test
-     * @return ScroreReport
-     */
-    ScoreReport score(Set<Incident> baseline, Set<Incident> sut);
-
-    /**
-     * The Name of the Strategy
-     */
-    String getName();
-
-}
