--- engine/driver/src/main/java/org/opennms/oce/engine/score/impl/ScoreReportBean.java
@@ -1,72 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2018 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2018 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.oce.engine.score.impl;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import org.opennms.oce.engine.score.api.ScoreReport;
-
-public class ScoreReportBean implements ScoreReport {
-    private double score;
-    private double maxScore = Double.POSITIVE_INFINITY;
-    private List<ScoreMetricBean> metrics = new ArrayList<>();
-
-    @Override
-    public double getScore() {
-        return score;
-    }
-
-    public void setScore(double score) {
-        this.score = score;
-    }
-
-    @Override
-    public double getMaxScore() {
-        return maxScore;
-    }
-
-    public void setMaxScore(double maxScore) {
-        this.maxScore = maxScore;
-    }
-
-    @Override
-    public List<ScoreMetricBean> getMetrics() {
-        return metrics;
-    }
-
-    public void setMetrics(List<ScoreMetricBean> metrics) {
-        this.metrics = metrics;
-    }
-
-    @Override
-    public String toString() {
-        return Double.toString(score) + metrics;
-    }
-}
