+++ engine/pom.xml
@@ -14,7 +14,6 @@
         <module>api</module>
         <module>cluster</module>
         <module>common</module>
-        <module>driver</module>
         <module>itest</module>
         <module>shell</module>
         <module>temporal</module>
